[DCP_1] Add solution 2 and library to C

+++ Daily_Coding_Problem/Problem_1/src/C/solution1.c
@@ -8,8 +8,8 @@ bool is_sum(uint64_t n, int64_t* arr, int64_t k)
         return false;
     }
 
-    for(int i = 0; i < n - 1; i++) {
-        for(int j = i + 1; j < n; j++) {
+    for(uint64_t i = 0; i < n - 1; i++) {
+        for(uint64_t j = i + 1; j < n; j++) {
             if(arr[i] + arr[j] == k) {
                 return true;
             }
